Split out all classes from organize_old.py
