+++ resqpy/organize/__init__.py
@@ -6,15 +6,22 @@ from .OrganizationFeature import OrganizationFeature
 from .GeobodyFeature import GeobodyFeature
 from .BoundaryFeature import BoundaryFeature
 
-# from .FrontierFeature import FrontierFeature
-# from .GeologicUnitFeature import GeologicUnitFeature
-# from .FluidBoundaryFeature import FluidBoundaryFeature
-# from .RockFluidUnitFeature import RockFluidUnitFeature
-# from .TectonicBoundaryFeature import TectonicBoundaryFeature
-# from .GeneticBoundaryFeature import GeneticBoundaryFeature
-# from .WellboreFeature import WellboreFeature
-# from .FaultInterpretation import FaultInterpretation
+from .FrontierFeature import FrontierFeature
+from .GeologicUnitFeature import GeologicUnitFeature
+from .FluidBoundaryFeature import FluidBoundaryFeature
+from .RockFluidUnitFeature import RockFluidUnitFeature
+from .TectonicBoundaryFeature import TectonicBoundaryFeature
+from .GeneticBoundaryFeature import GeneticBoundaryFeature
+from .WellboreFeature import WellboreFeature
+from .FaultInterpretation import FaultInterpretation
 from .EarthModelInterpretation import EarthModelInterpretation
-# from .HorizonInterpretation import HorizonInterpretation
-# from .GeobodyBoundaryInterpretation import GeobodyBoundaryInterpretation
-# from .GeobodyInterpretationllboreInterpretation import GeobodyInterpretationllboreInterpretation
+from .HorizonInterpretation import HorizonInterpretation
+from .GeobodyBoundaryInterpretation import GeobodyBoundaryInterpretation
+from .GeobodyInterpretation import GeobodyInterpretation
+from .WellboreInterpretation import WellboreInterpretation
+
+from resqpy.organize.organize_functions import (
+    equivalent_extra_metadata,
+    alias_for_attribute,
+    extract_has_occurred_during, equivalent_chrono_pairs, create_xml_has_occurred_during
+)
