+++ resqpy/well.py
@@ -52,7 +52,8 @@ import resqpy.olio.vector_utilities as vec
 import resqpy.olio.wellspec_keywords as wsk
 import resqpy.olio.write_hdf5 as rwh5
 import resqpy.olio.xml_et as rqet
-import resqpy.organize.organize_old as rqo
+# import resqpy.organize.organize_old as rqo
+import resqpy.organize as rqo
 import resqpy.property as rqp
 import resqpy.weights_and_measures as bwam
 from resqpy.olio.base import BaseResqpy
