+++ tests/conftest.py
@@ -13,7 +13,8 @@ import resqpy.olio.xml_et as rqet
 import resqpy.property as rqp
 from resqpy.crs import Crs
 from resqpy.model import Model, new_model
-from resqpy.organize.organize_old import WellboreFeature, WellboreInterpretation
+# from resqpy.organize.organize_old import WellboreFeature, WellboreInterpretation
+from resqpy.organize import WellboreFeature, WellboreInterpretation
 from resqpy.well import MdDatum, Trajectory, WellboreFrame
